Update docker-compose.yml (#148)
* Update docker-compose.yml

In Celery 5.0, -A or --app should come first, as a global option. Running celery worker -A results in the following output:

"You are using `-A` as an option of the worker sub-command:
celery worker -A celeryapp <...>

The support for this usage was removed in Celery 5.0. Instead you should use `-A` as a global option:
celery -A celeryapp worker <...>"

* Update docker-compose.yml